get_secondary_factor now allows for any double value, and uses a math function, and the struct node allows for a double

+++ src/graph_functions.h
@@ -6,7 +6,7 @@
 
 typedef struct {
     int node_id;
-    int quality_of_stop;
+    double quality_of_stop;
     int charger_present;
 } node_t;
